test_SB/requirements.txt
Django==3.2
djangorestframework==3.11.1
DateTime==4.3
gunicorn==20.1.0
python-memcached==1.59
psycopg2==2.9.5
python-dotenv==0.21.0